+++ nosql_schema/db/nosqlite/collection_handler.py
@@ -69,11 +69,12 @@ class CollectionHandler(AbstractCollectionHandler):
         return self.collection_handle.count(query)
 
     @staticmethod
-    def convert_ids(query):
+    def convert_ids(query, is_id=False):
         """
         Casts all '_id' to 'int'
 
         :param query: Query that has to be transformed
+        :param is_id: Indicates that the current part derived from an '_id'
         :return: Transformed query (dict)
         """
         if not isinstance(query, dict):
@@ -82,8 +83,10 @@ class CollectionHandler(AbstractCollectionHandler):
         for k,v in query.iteritems():
             if k == '_id':
                 if isinstance(v, dict):
-                    query[k] = CollectionHandler.convert_ids(v)
+                    query[k] = CollectionHandler.convert_ids(v, True)
                 elif type(v) in [str, unicode]:
                     query[k] = int(v)
+            elif is_id and type(v) in [str, unicode]:
+                query[k] = int(v)
 
         return query