+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Fixed
 - Fixed `drop_collection` on MongoDB
+- Fixed `convert_ids` to handle operators on `_id`
 
 ## [0.1.0] - 2016-08-12
 ### Added