+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [0.2.0] - [unreleased]
+## [0.2.0] - 2016-09-26
 ### Added
 - Added possibility to create an object by dictionary
 (pass `__dictionary` argument)